Create Blacklist Non nvidia vulkan drivers
This is what you NEED to do if you get the same bullshit I have been getting with my proton logs. You have to disable Vulkan from trying to use the mesa drivers. If there was just a option for you to Select what proton driver you want to use maybe in the steam compatibility options menu so YOU pick which driver you want to use for steam it would be 10 times better. So i don't have to go thru this BS again